GP-37735: Refactor adding 'venue contact person realationship' block to the case summary.
